Splittes booking service to separate services
