+++ Api/Services/Accommodations/Bookings/IBookingManagementService.cs
@@ -11,16 +11,8 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings
 {
-    public interface IBookingService
+    public interface IBookingManagementService
     {
-        Task<Result<string, ProblemDetails>> Register(AccommodationBookingRequest bookingRequest, AgentContext agent, string languageCode);
-
-        Task<Result<AccommodationBookingInfo, ProblemDetails>> Finalize(string referenceCode, AgentContext agentContext, string languageCode);
-
-        Task<Result<AccommodationBookingInfo, ProblemDetails>> BookByAccount(AccommodationBookingRequest bookingRequest, AgentContext agentContext, string languageCode, string clientIp);
-        
-        Task ProcessResponse(Booking bookingResponse, Data.Booking.Booking booking);
-
         Task<Result<VoidObject, ProblemDetails>> Cancel(int bookingId, AgentContext agent);
         
         Task<Result<VoidObject, ProblemDetails>> Cancel(int bookingId, ServiceAccount serviceAccount);
