+++ HappyTravel.Edo.UnitTests/Tests/Services/Accommodations/Bookings/BookingsProcessingServiceTests/Cancellation/ExecutingCancellation.cs
@@ -19,7 +19,7 @@ namespace HappyTravel.Edo.UnitTests.Tests.Services.Accommodations.Bookings.Booki
         [Fact]
         public async Task Cancel_valid_bookings_should_succeed()
         {
-            var service = CreateProcessingService(Mock.Of<IBookingService>());
+            var service = CreateProcessingService(Mock.Of<IBookingManagementService>());
 
             var (isSuccess, _, _, _) = await service.Cancel(new List<int> {1, 2}, new ServiceAccount() {ClientId = "ClientId", Id = 12});
 
@@ -30,7 +30,7 @@ namespace HappyTravel.Edo.UnitTests.Tests.Services.Accommodations.Bookings.Booki
         [Fact]
         public async Task Cancel_invalid_booking_should_fail()
         {
-            var service = CreateProcessingService(Mock.Of<IBookingService>());
+            var service = CreateProcessingService(Mock.Of<IBookingManagementService>());
 
             var (_, isFailure, _, _) = await service.Cancel(new List<int> {3, 4}, new ServiceAccount() {ClientId = "ClientId", Id = 12});
 
@@ -41,7 +41,7 @@ namespace HappyTravel.Edo.UnitTests.Tests.Services.Accommodations.Bookings.Booki
         [Fact]
         public async Task All_bookings_should_be_cancelled()
         {
-            var bookingServiceMock = new Mock<IBookingService>();
+            var bookingServiceMock = new Mock<IBookingManagementService>();
             var service = CreateProcessingService(bookingServiceMock.Object);
 
             await service.Cancel(new List<int> {1, 2}, new ServiceAccount() {ClientId = "ClientId", Id = 12});
@@ -54,14 +54,14 @@ namespace HappyTravel.Edo.UnitTests.Tests.Services.Accommodations.Bookings.Booki
         }
 
 
-        private BookingsProcessingService CreateProcessingService(IBookingService bookingService)
+        private BookingsProcessingService CreateProcessingService(IBookingManagementService bookingManagementService)
         {
             var context = MockEdoContextFactory.Create();
             context.Setup(c => c.Bookings)
                 .Returns(DbSetMockProvider.GetDbSetMock(Bookings));
 
             var service = new BookingsProcessingService(Mock.Of<IBookingPaymentService>(), 
-               bookingService, 
+               bookingManagementService, 
                Mock.Of<IBookingMailingService>(),
                 context.Object);
             return service;
