+++ HappyTravel.Edo.UnitTests/Tests/Services/Accommodations/Bookings/BookingsProcessingServiceTests/Capturing/GettingForCapturing.cs
@@ -192,7 +192,7 @@ namespace HappyTravel.Edo.UnitTests.Tests.Services.Accommodations.Bookings.Booki
                 .Returns(DbSetMockProvider.GetDbSetMock(bookings));
 
             var service = new BookingsProcessingService(Mock.Of<IBookingPaymentService>(),
-                Mock.Of<IBookingService>(),
+                Mock.Of<IBookingManagementService>(),
                 Mock.Of<IBookingMailingService>(),
                 context.Object);
